+++ spec/features/school_buying_professionals/view_a_list_of_tasks_spec.rb
@@ -180,15 +180,15 @@ feature "Users can view the task list" do
         click_on("One additional question task")
         click_on(I18n.t("generic.button.back"))
 
-        steps = find_all(".app-task-list__item.step__item")
-        within(".app-task-list") do
+        steps = find_all(".govuk-summary-list__row")
+        within(".govuk-summary-list") do
           expect(steps[0]).to have_content("What support do you have available?")
           # Hidden field "What colour is the sky?" is correctly omitted at this point
           expect(steps[1]).to have_content("What email address did you use?")
         end
 
         # Answer the first question to unlock "What colour is the sky?"
-        click_on("What support do you have available?")
+        first(".govuk-summary-list__row").click_on(I18n.t("generic.button.change_answer"))
         choose("School expert")
         click_on("Continue")
 
@@ -196,8 +196,8 @@ feature "Users can view the task list" do
         click_on(I18n.t("generic.button.back"))
 
         # Check that "What colour is in the sky added to the correct place in the list"
-        steps = find_all(".app-task-list__item.step__item")
-        within(".app-task-list") do
+        steps = find_all(".govuk-summary-list__row")
+        within(".govuk-summary-list") do
           expect(steps[0]).to have_content("What support do you have available?")
           expect(steps[1]).to have_content("What colour is the sky?")
           expect(steps[2]).to have_content("What email address did you use?")